Dedup LoadState and DataLoadState interfaces b/w webapp/core and debugger_v2
