+++ tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/store/BUILD
@@ -23,6 +23,7 @@ ng_module(
         "debugger_types.ts",
     ],
     deps = [
+        "//tensorboard/webapp/types",
         "@npm//@ngrx/store",
         "@npm//rxjs",
     ],
