+++ tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/testing/index.ts
@@ -17,8 +17,8 @@ import {Component, NgModule} from '@angular/core';
 import {Store} from '@ngrx/store';
 
 import {
-  DEBUGGER_FEATURE_KEY,
   DataLoadState,
+  DEBUGGER_FEATURE_KEY,
   DebuggerState,
   State,
 } from '../store/debugger_types';
