+++ tensorboard/webapp/core/effects/core_effects.ts
@@ -32,7 +32,7 @@ import {
   pluginsListingFailed,
 } from '../actions';
 import {getPluginsListLoaded} from '../store';
-import {LoadState} from '../../types/api';
+import {DataLoadState} from '../../types/api';
 import {State} from '../store/core_types';
 import {TBServerDataSource} from '../../webapp_data_source/tb_server_data_source';
 
@@ -50,7 +50,7 @@ export class CoreEffects {
     this.actions$.pipe(
       ofType(coreLoaded, reload),
       withLatestFrom(this.store.select(getPluginsListLoaded)),
-      filter(([, {state}]) => state !== LoadState.LOADING),
+      filter(([, {state}]) => state !== DataLoadState.LOADING),
       tap(() => this.store.dispatch(pluginsListingRequested())),
       mergeMap(() => {
         return zip(
