+++ tensorboard/webapp/core/effects/core_effects_test.ts
@@ -28,7 +28,7 @@ import {State} from '../store';
 
 import {createPluginMetadata, createState, createCoreState} from '../testing';
 
-import {PluginsListing, LoadState as DataLoadState} from '../../types/api';
+import {PluginsListing, DataLoadState} from '../../types/api';
 import {TBServerDataSource} from '../../webapp_data_source/tb_server_data_source';
 
 describe('core_effects', () => {
