+++ tensorboard/webapp/core/store/core_selectors.ts
@@ -14,8 +14,8 @@ limitations under the License.
 ==============================================================================*/
 
 import {createSelector, createFeatureSelector} from '@ngrx/store';
-import {PluginId, PluginsListing} from '../../types/api';
-import {CoreState, State, CORE_FEATURE_KEY, LoadState} from './core_types';
+import {LoadState, PluginId, PluginsListing} from '../../types/api';
+import {CoreState, State, CORE_FEATURE_KEY} from './core_types';
 
 // HACK: These imports are for type inference.
 // https://github.com/bazelbuild/rules_nodejs/issues/1013
