+++ tensorboard/webapp/types/api.ts
@@ -12,11 +12,18 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
-export enum LoadState {
+
+export enum DataLoadState {
   NOT_LOADED,
   LOADED,
   LOADING,
   FAILED,
+}
+
+export interface LoadState {
+  state: DataLoadState;
+  // Time since epoch.
+  lastLoadedTimeInMs: number | null;
 }
 
 export type PluginId = string;
